requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
alembic==1.12.1
python-multipart==0.0.6
click==8.1.7
pandas==2.1.3
prometheus-client==0.19.0
nats-py==2.2.0
psutil==5.9.5
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0